requirements.txt
accelerate==0.23.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.2
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.3.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bce-python-sdk==0.8.90
bcrypt==4.1.2
beautifulsoup4==4.12.2
bleach==6.0.0
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
coloredlogs==15.0.1
colorlog==6.7.0
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.3
datasets==2.14.5
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.21.1
dill==0.3.7
erniebot==0.3.1
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.103.2
fastjsonschema==2.18.0
ffmpy==0.3.1
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
google-auth==2.26.1
googleapis-common-protos==1.62.0
gradio==3.47.1
gradio_client==0.6.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface==0.0.1
huggingface-hub==0.17.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.1
ipython==8.15.0
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.353
langchain-community==0.0.7
langchain-core==0.1.4
langsmith==0.0.77
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mistune==3.0.1
mmh3==4.0.1
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.25.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==0.28.1
openapi-core==0.18.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.6.0
opencv-python==4.8.0.76
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.8
overrides==7.4.0
packaging==23.2
pandas==2.1.0
pandocfilters==1.5.0
parse==1.19.1
parso==0.8.3
pathable==0.4.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
posthog==3.1.0
prance==23.6.21.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.0
rsa==4.9
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safetensors==0.3.3
semantic-kernel==0.3.13.dev0
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
terminado==0.17.1
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==3.0.0
wrapt==1.16.0
xxhash==3.3.0
yarl==1.9.2
zhipuai==1.0.7
zipp==3.16.2
